unconvert: drop 12 redundant type conversions
Every flagged conversion casts a value to the type it already has —
no behaviour change, just dead noise. Concretely:

- pkg/io_uring/bench_test.go: int64(ru.Utime.Usec) — Usec is already
  int64 on linux/amd64 (syscall.Timeval).
- pkg/xtcp/deserialize.go: uint64(x.config.Modulus) — Modulus is uint64.
- pkg/xtcp/init.go: uint32(x.config.NlmsgSeq) — NlmsgSeq is uint32.
- pkg/xtcpnl/xtcpnl.go: uint32(r.NlMsgLen), uint32(r.NlMsgSeq),
  uint8(r.AddressFamily), uint8(r.IDiagExt) — all already match.
- pkg/xtcpnl/xtcpnl_inet_diag_tcpinfo.go:932: uint32(binary.LittleEndian.Uint32(...))
  — Uint32 already returns uint32.
- pkg/xtcpnl/xtcpnl_serialize_netlink_request.go: byte(...) on uint8 fields.

Build pinned to linux/amd64 via nix/microvms/constants.nix, so the
linux/386 `Timeval{Sec int32, Usec int32}` portability hedge isn't
relevant here.

+++ pkg/xtcp/deserialize.go
@@ -75,7 +75,7 @@ func (x *XTCP) Deserialize(ctx context.Context, d DeserializeArgs) (n uint64, er
 		// }
 
 		if x.config.Modulus != 1 {
-			if n%uint64(x.config.Modulus) != 1 {
+			if n%x.config.Modulus != 1 {
 				d.pC.WithLabelValues("Deserialize", "continue", "count").Inc()
 				continue
 			}
